Stop docker-compose errors when removing services
When removing a service (e.g. service A) from the environment
configuration, `navy refresh-config` would fail because we were calling
`docker-compose up` with a list of service names, including service A.

This change makes navy check the list of service names it's trying to
launch, and filters out any that it doesn't know about.