Improve navy update when no services are running
Currently, if you run 'navy update' when no services are running, it
will update all images, but also launch all services, which is probably
not what you want.

What this now does in that case is to update all images, and launch no
services.

In general, 'navy update' will now:

- navy update:
  - update all images
  - launch all currently running services
- navy update <services>
  - update given images
  - launch given images, only if already running