Fix integration testing
The integration tests run in their own container and navy
uses a dedicated docker daemon running in another container
on the host VM. We do not pass the environment variables
from the travis matrix down to the scripts that setup this
testing environment. Hence until now, all test builds have
been using the same version of docker and docker-compose.