getLaunchedNavies now validates that a navy path is a directory
Calls to Navy.safeGetDriver within various methods were throwing
an Error when that navies directory contained a file. One example
being the unintended existence of a .DS_Store file on MacOS.
getLaunchedNavies now filters out all non-directory nodes. Manually
creating an empty directory instead would still surface the problem,
however this would be a seen as an intentional corruption of the navies
directory and the doctor command can be used to resolve this.

Fixes #68