Remove _cachedState
It was causing problems where config provided in `initialise` was saved
into the state.json file but was not making it into the
docker-compose.yml.

Could also be solved by making `initialise` call `this.saveState` rather
than calling `saveState` directly or invalidating the cached state in
`destroy`.

There seems to be no speed difference without the caching so let's try
without.